[add] Support inline assembly (__asm__) and fix codegen bugs
- Fixed code generation for function calls with operators like ND_SUB, ND_POSTINCREMENT, etc.
- Previously, the RDI and RSI registers were corrupted in these cases.

+++ include/eval_constant.h
@@ -1,6 +1,6 @@
 #ifndef EVAL_CONSTANT_C_COMPILER
 #define EVAL_CONSTANT_C_COMPILER
 
-long eval_constant_expression();
+long long eval_constant_expression();
 
 #endif
